requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
attrs==20.3.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-celery==3.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.11.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==3.0.37
MarkupSafe==1.1.1
numpy==1.19.4
openapi-codec==1.3.2
packaging==20.7
pandas==1.1.5
pathlib==1.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
